backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-RESTful==0.3.10
Flask-CORS==4.0.0
python-dotenv==1.0.0
marshmallow==3.20.1
cloudinary==1.36.0
Pillow==10.1.0
Werkzeug==2.3.7
